requirements.txt
django
djangorestframework
requests
psycopg2-binary
beautifulsoup4
django-cors-headers
